requirements.txt
openai
google-generativeai
anthropic
tqdm
Pillow
python-dotenv